feat: here comes streamable logger
