+++ rsp-libs/src/logger/console_logger.cpp
@@ -5,60 +5,15 @@ namespace rsp {
 namespace libs {
 namespace logger {
 
-s_logger& s_logger::operator<<(flags flag) {
-  if (is_null()) return *this;
-  switch (flag) {
-    case L_time:
-      log_time();
-      break;
-    case L_flush:
-      _flags = static_cast<flags>(_flags & L_allwaysFlush);
-      *this << " |F|\n";
-      flush();
-      break;
-    case L_endl: {
-      if (_flags & L_allwaysFlush) {
-        *this << " |F|";
-      } else if (_flags == L_startWithFlushing) {
-        *this << " |SF|";
-      }
-      auto streamPtr = &stream();
-      s_logger* logger = this;
-      do {
-        *streamPtr << "\n";
-        logger = logger->mirror_stream(streamPtr);
-      } while (streamPtr);
-      if (_flags & L_allwaysFlush || _flags == L_startWithFlushing) flush();
-    }
-      [[fallthrough]];
-    case L_clearflags:
-      if (_flags != L_startWithFlushing) {
-        _flags = static_cast<flags>(_flags & L_allwaysFlush);
-      }
-      break;
-    case L_allwaysFlush:
-      _flags += L_allwaysFlush;
-      break;
-    case L_concat:
-      remove_flag(L_tabs);
-      break;
-    default:
-      add_flag(flag);
-  }
-  return *this;
-}
-tm* s_logger::get_time() {
-  std::time_t now = std::time(nullptr);
-  auto localTime = std::localtime(&now);
-  log_date.dayNo = localTime->tm_mday;
-  log_date.monthNo = localTime->tm_mon + 1;
-  return localTime;
-}
+std::once_flag console_logger::s_flag;
+std::unique_ptr<console_logger> console_logger::s_instance;
+std::once_flag console_logger::s_mirror_flag;
 
-s_logger& s_logger::log_time() {
-  *this << std::put_time(get_time(), "%d/%m/%y %H:%M:%S");
-  _flags += L_time;
-  return *this;
+console_logger& console_logger::instance() {
+  std::call_once(console_logger::s_flag, []() {
+    console_logger::s_instance.reset(new console_logger{log_level::TRACE});
+  });
+  return *console_logger::s_instance;
 }
 
 }  // namespace logger
