+++ rsp-libs/src/logger/streamable_logger.cpp
@@ -4,19 +4,60 @@ namespace rsp {
 namespace libs {
 namespace logger {
 
-template <typename T>
-s_logger& s_logger::log(T value) {
+s_logger& s_logger::operator<<(flags flag) {
   if (is_null()) return *this;
-  auto streamPtr = &stream();
-  s_logger* logger = this;
-  do {
-    if (is_tabs()) {
-      *streamPtr << "\t";
+  switch (flag) {
+    case L_time:
+      log_time();
+      break;
+    case L_flush:
+      _flags = static_cast<flags>(_flags & L_allwaysFlush);
+      *this << " |F|\n";
+      flush();
+      break;
+    case L_endl: {
+      if (_flags & L_allwaysFlush) {
+        *this << " |F|";
+      } else if (_flags == L_startWithFlushing) {
+        *this << " |SF|";
+      }
+      auto stream_ptr = &stream();
+      s_logger* logger = this;
+      do {
+        *stream_ptr << "\n";
+        logger = logger->mirror_stream(&stream_ptr);
+      } while (stream_ptr);
+      if (_flags & L_allwaysFlush || _flags == L_startWithFlushing) flush();
     }
-    *streamPtr << value;
-    logger = logger->mirror_stream(streamPtr);
-  } while (streamPtr);
-  remove_flag(L_time);
+      [[fallthrough]];
+    case L_clearflags:
+      if (_flags != L_startWithFlushing) {
+        _flags = static_cast<flags>(_flags & L_allwaysFlush);
+      }
+      break;
+    case L_allwaysFlush:
+      _flags += L_allwaysFlush;
+      break;
+    case L_concat:
+      remove_flag(L_tabs);
+      break;
+    default:
+      add_flag(flag);
+  }
+  return *this;
+}
+
+tm* s_logger::get_time() {
+  std::time_t now = std::time(nullptr);
+  auto localTime = std::localtime(&now);
+  log_date.dayNo = localTime->tm_mday;
+  log_date.monthNo = localTime->tm_mon + 1;
+  return localTime;
+}
+
+s_logger& s_logger::log_time() {
+  *this << std::put_time(get_time(), "%d/%m/%y %H:%M:%S");
+  _flags += L_time;
   return *this;
 }
